Support v2 of Smugmug API
Usage:

```js
  smugmug: {
    "resource": smugmug.v2["user!albums"],
    "params": {
      "NickName": "sanjacinto-museum",
      "APIKey": "fSMJ7g7ju5aZgYsNiMl3PstAQYzmbg9L",
      "Order": "Album+Settings"
      // "count": 7
    },
    "auth": {
      "type": "oauth",
      "provider": "smugmug"
    },
    "cleanup": function(data) {
       // `data` contains "Albums" and "Total" keys.
    }
  }
```

Smugmug API v2 supports a "_config" object to effectively perform 
actions to multiple endpoints within a single request. Theoretically,
I could get all the data in one request. However, their server did
not handle that much, and the request timeout'ed.

Therefore, I have to split the request into pages. Except the first
request, which tells us the total albums and the number of albums
per page, all other requests are performed in parallel.

The number of requests is:

`1 + N/X`, in which N is the total number of albums, and X is the
amount of albums per page, usually 50.